Refactoring SRP를 위해 AuthUserWorkerService와 WorkerRegistrationService를 나눔
