+++ src/main/java/site/hirecruit/hr/domain/worker/service/AuthUserWorkerService.kt
@@ -12,13 +12,6 @@ import site.hirecruit.hr.domain.worker.dto.WorkerDto
 interface AuthUserWorkerService {
 
     /**
-     * [AuthUserInfo]와 [WorkerDto.Registration]를 기반으로 [WorkerEntity]를 생성 및 저장한다.
-     *
-     * @return Worker의 정보
-     */
-    fun registration(authUserInfo: AuthUserInfo, registrationDto: WorkerDto.Registration): WorkerDto.Info
-
-    /**
      * [AuthUserInfo]를 기반으로 Worker에 대한 정보를 찾는다.
      *
      * @return Worker의 정보
